Extract Haste tracking to its own module
