+++ src/Parser/HolyPaladin/Modules/Features/AlwaysBeCasting.js
@@ -49,8 +49,6 @@ class AlwaysBeCasting extends CoreAlwaysBeCastingHealing {
   ];
 
   on_initialized() {
-    super.on_initialized();
-
     const combatant = this.combatants.selected;
 
     if (combatant.hasTalent(SPELLS.CRUSADERS_MIGHT_TALENT.id)) {
